backend/requirements.txt
# 🛒 沃尔玛AI Agent平台 - Python依赖包
# Walmart AI Agent Platform - Python Dependencies

# ===== 核心框架 =====
fastapi==0.104.1              # 高性能Web框架
uvicorn[standard]==0.24.0     # ASGI服务器
pydantic==2.5.0               # 数据验证和序列化
python-multipart==0.0.6       # 文件上传支持

# ===== 数据库相关 =====
sqlalchemy==2.0.23            # ORM框架
alembic==1.13.0               # 数据库迁移
asyncpg==0.29.0               # PostgreSQL异步驱动
psycopg2-binary==2.9.9        # PostgreSQL同步驱动
redis==5.0.1                  # Redis客户端

# ===== 向量数据库 =====
chromadb==0.4.18              # 向量数据库
pinecone-client==2.2.4        # Pinecone向量数据库
faiss-cpu==1.7.4              # Facebook向量搜索库

# ===== AI/ML相关 =====
langchain==0.0.350            # LLM应用框架
langchain-community==0.0.3    # LangChain社区扩展
openai==1.3.7                 # OpenAI API
anthropic==0.7.7              # Claude API
transformers==4.36.0          # Hugging Face模型库
torch==2.1.1                  # PyTorch深度学习框架
sentence-transformers==2.2.2  # 句子嵌入模型
tiktoken==0.5.2               # OpenAI分词器

# ===== Dify集成 =====
dify-client==0.1.5            # Dify官方SDK
requests==2.31.0              # HTTP客户端
httpx==0.25.2                 # 异步HTTP客户端

# ===== 数据处理 =====
pandas==2.1.4                 # 数据分析库
numpy==1.24.4                 # 数值计算库
scikit-learn==1.3.2           # 机器学习库
nltk==3.8.1                   # 自然语言处理
jieba==0.42.1                 # 中文分词
openpyxl==3.1.2               # Excel文件处理
python-docx==1.1.0            # Word文档处理
PyPDF2==3.0.1                 # PDF文件处理
python-pptx==0.6.23           # PPT文件处理

# ===== 异步任务 =====
celery==5.3.4                 # 分布式任务队列
flower==2.0.1                 # Celery监控工具

# ===== 认证授权 =====
python-jose[cryptography]==3.3.0  # JWT处理
passlib[bcrypt]==1.7.4        # 密码加密
python-multipart==0.0.6       # 表单数据处理

# ===== 配置管理 =====
python-dotenv==1.0.0          # 环境变量管理
pydantic-settings==2.1.0      # 配置管理

# ===== 监控日志 =====
loguru==0.7.2                 # 高级日志库
prometheus-client==0.19.0     # Prometheus监控
structlog==23.2.0             # 结构化日志

# ===== 工具库 =====
typer==0.9.0                  # CLI工具
rich==13.7.0                  # 美化终端输出
python-dateutil==2.8.2       # 日期时间处理
pytz==2023.3.post1            # 时区处理
shortuuid==1.0.11             # 短UUID生成
python-magic==0.4.27          # 文件类型检测

# ===== 测试相关 =====
pytest==7.4.3                 # 测试框架
pytest-asyncio==0.21.1        # 异步测试支持
pytest-cov==4.1.0             # 测试覆盖率
httpx==0.25.2                 # 测试HTTP客户端
factory-boy==3.3.0            # 测试数据工厂

# ===== 开发工具 =====
black==23.11.0                # 代码格式化
flake8==6.1.0                 # 代码检查
mypy==1.7.1                   # 类型检查
pre-commit==3.6.0             # Git钩子

# ===== 部署相关 =====
gunicorn==21.2.0              # WSGI服务器
supervisor==4.2.5             # 进程管理

# ===== 文档生成 =====
mkdocs==1.5.3                 # 文档生成工具
mkdocs-material==9.4.8        # Material主题
